refactor: Update import paths to use absolute paths and improve project structure
