+++ .swcrc
@@ -1,0 +1,23 @@
+{
+  "jsc": {
+    "parser": {
+      "syntax": "typescript",
+      "tsx": false,
+      "dynamicImport": true,
+      "decorators": false
+    },
+    "target": "esnext",
+    "baseUrl": "./",
+    "paths": {
+      "@/*": ["./src/*"],
+      "@languages/*": ["./languages/*"]
+    }
+  },
+  "module": {
+    "type": "commonjs",
+    "strictMode": false,
+    "lazy": true
+  },
+  "minify": true,
+  "sourceMaps": true
+}
