+++ package.json
@@ -1,13 +1,13 @@
 {
   "name": "botrucho",
-  "version": "1.2.0",
+  "version": "1.3.0",
   "description": "",
   "main": "src/sharder.ts",
   "scripts": {
-    "clean": "rmdir /s /q dist",
-    "copy-assets": "npx cpx 'assets/**/*' dist/assets",
-    "copy-env": "npx cpx '.env' dist",
-    "build": "tsc && tsc-alias && npm run copy-assets && npm run copy-env",
+    "clean": "rm -rf dist",
+    "copy-env": "npx cpx \".env\" dist/",
+    "copy-assets": "npx cpx \"assets/**/*\" dist/assets && npx cpx \"languages/**/*\" dist/languages",
+    "build": "swc src -d dist --copy-files && npm run copy-assets && npm run copy-env",
     "start": "npm run build && node dist/src/sharder.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -24,7 +24,6 @@
     "discord-player-youtubei": "^1.4.6",
     "discord-voip": "^7.1.0",
     "discord.js": "^14.19.3",
-    "dotenv": "^16.5.0",
     "express": "^5.1.0",
     "jimp": "^1.6.0",
     "microsoft-cognitiveservices-speech-sdk": "^1.44.0",
@@ -32,9 +31,10 @@
     "node-cron": "^4.0.6"
   },
   "devDependencies": {
+    "@swc/cli": "^0.7.7",
+    "@swc/core": "^1.11.29",
     "@types/express": "^5.0.2",
     "@types/node-cron": "^3.0.11",
-    "tsc-alias": "^1.8.16",
     "typescript": "^5.8.2"
   }
 }
