+++ src/commands/music/skip.ts
@@ -1,11 +1,11 @@
 import { CommandInteraction, EmbedBuilder, SlashCommandOptionsOnlyBuilder } from "discord.js";
 import { SlashCommandBuilder } from "@discordjs/builders";
 import { GuildQueue, useQueue } from "discord-player";
-import Botrucho from "@mongodb/base/Botrucho";
-import { IGuildData } from "@mongodb/models/GuildData";
-import { Error, Success } from "@util/embedMessage";
-import { MusicKeys, SkipKeys, TranslationElement } from "@customTypes/Translations";
-import { PlayerMetadata } from "@customTypes/PlayerMetadata";
+import Botrucho from "@/mongodb/base/Botrucho";
+import { IGuildData } from "@/mongodb/models/GuildData";
+import { PlayerMetadata } from "@/types/PlayerMetadata";
+import { MusicKeys, SkipKeys, TranslationElement } from "@/types/Translations";
+import { Error, Success } from "@/util/embedMessage";
 
 export const name = 'skip';
 export const description = 'Skip to the next track.';
