+++ src/events/discord/messageDelete.ts
@@ -1,5 +1,5 @@
 import { Message } from "discord.js";
-import Botrucho from "@mongodb/base/Botrucho";
+import Botrucho from "@/mongodb/base/Botrucho";
 
 export async function execute(_: any, message: Message & { client: Botrucho }) {
   const { client: { deleted_messages } } = message;
