+++ src/events/player/playerError.ts
@@ -1,4 +1,4 @@
-import { logger } from '@util/Logger';
+import { logger } from '@/util/Logger';
 
 export async function execute(error: any) {
   logger.error("[ERROR]", error)
