+++ src/events/player/volumeChange.ts
@@ -1,6 +1,6 @@
-import Botrucho from "@mongodb/base/Botrucho";
 import { GuildQueue } from "discord-player";
-import { PlayerMetadata } from "@customTypes/PlayerMetadata";
+import Botrucho from "@/mongodb/base/Botrucho";
+import { PlayerMetadata } from "@/types/PlayerMetadata";
 
 export async function execute(client: Botrucho, queue: GuildQueue<PlayerMetadata>, _oldVolume: number, newVolume: number): Promise<void> {
   const { guildData } = client;
