+++ src/util/Logger.ts
@@ -1,6 +1,31 @@
 import { Tint } from './logColors';
 
 class Logger {
+  static log(...messages: any[]): void {
+    const combinedMessage = messages.map(msg => (typeof msg === 'object' ? JSON.stringify(msg) : msg)).join(' ');
+    console.log(this.formatLog(Tint.green("LOG"), Tint.green(combinedMessage)));
+  }
+
+  static warn(...messages: any[]): void {
+    const combinedMessage = messages.map(msg => (typeof msg === 'object' ? JSON.stringify(msg) : msg)).join(' ');
+    console.log(this.formatLog(Tint.yellow("WARN"), Tint.yellow(combinedMessage)));
+  }
+
+  static error(...messages: any[]): void {
+    const combinedMessage = messages.map(msg => (typeof msg === 'object' ? JSON.stringify(msg) : msg)).join(' ');
+    console.log(this.formatLog(Tint.red("ERROR"), Tint.red(combinedMessage)));
+  }
+
+  static debug(...messages: any[]): void {
+    const combinedMessage = messages.map(msg => (typeof msg === 'object' ? JSON.stringify(msg) : msg)).join(' ');
+    console.log(this.formatLog(Tint.cyan("DEBUG"), Tint.cyan(combinedMessage)));
+  }
+
+  static trace(...messages: any[]): void {
+    const combinedMessage = messages.map(msg => (typeof msg === 'object' ? JSON.stringify(msg) : msg)).join(' ');
+    console.log(this.formatLog(Tint.blue("TRACE"), Tint.blue(combinedMessage)));
+  }
+
   private static getTimestamp(): string {
     return new Date().toLocaleString('en-GB', {
       hour12: true,
@@ -28,31 +53,6 @@ class Logger {
     const timestamp = this.getTimestamp();
     const caller = this.getCaller();
     return `${Tint.magenta(`[Botrucho] ${pid.toString().padEnd(6)} -`)} ${Tint.white(timestamp)}      ${level} ${Tint.magenta(`[${caller}]`)} ${message}`;
-  }
-
-  static log(...messages: any[]): void {
-    const combinedMessage = messages.map(msg => (typeof msg === 'object' ? JSON.stringify(msg) : msg)).join(' ');
-    console.log(this.formatLog(Tint.green("LOG"), Tint.green(combinedMessage)));
-  }
-
-  static warn(...messages: any[]): void {
-    const combinedMessage = messages.map(msg => (typeof msg === 'object' ? JSON.stringify(msg) : msg)).join(' ');
-    console.log(this.formatLog(Tint.yellow("WARN"), Tint.green(combinedMessage)));
-  }
-
-  static error(...messages: any[]): void {
-    const combinedMessage = messages.map(msg => (typeof msg === 'object' ? JSON.stringify(msg) : msg)).join(' ');
-    console.log(this.formatLog(Tint.red("ERROR"), Tint.green(combinedMessage)));
-  }
-
-  static debug(...messages: any[]): void {
-    const combinedMessage = messages.map(msg => (typeof msg === 'object' ? JSON.stringify(msg) : msg)).join(' ');
-    console.log(this.formatLog(Tint.cyan("DEBUG"), Tint.green(combinedMessage)));
-  }
-
-  static trace(...messages: any[]): void {
-    const combinedMessage = messages.map(msg => (typeof msg === 'object' ? JSON.stringify(msg) : msg)).join(' ');
-    console.log(this.formatLog(Tint.blue("TRACE"), Tint.green(combinedMessage)));
   }
 }
